(Java) Use a single, shared Java global scope across files
This conforms to the architecture of the Java language. Unlike
C/C++, each file does not have its own global scope. In Java,
there is only one global scope, which is then shared amongst
all the files.

Now, there are separate Packages, which can be viewed as a level
of scoping.

We still maintain each SgFile's SgGlobal* globalVariable because
this maintains symmetry within ROSE, and also provides us with
a "normal" way to store each file's information (file info,
declarations, etc.) for traversal.